- Removed initrd parameter from readme - Added function to set device depending on parameter - Reworked set flag function to set general cflags - Reworked configure.ac a bit - Added defines for no sanitize, no return and constructor - Replaced mmio_write by io_out{8/16/32} - Replaced mmio_read by io_in{8/16/32} - Splitted up avl functions into multiple files - Splitted tar functions into multiple files
