+++ include/lib/assert.h
@@ -23,7 +23,7 @@
 
 #include <stdint.h>
 
-void _assert( const char* restrict, uint32_t, const char* restrict ) __attribute__((noreturn));
+void __no_return _assert( const char* restrict, uint32_t, const char* restrict );
 
 #define assert( b ) ( b ? ( void )0 : _assert( __FILE__, __LINE__, #b ) );
