+++ include/lib/avl.h
@@ -59,4 +59,6 @@ void avl_remove_by_data( const avl_tree_ptr_t, void* );
 void avl_insert_by_node( const avl_tree_ptr_t, avl_node_ptr_t );
 void avl_remove_by_node( const avl_tree_ptr_t, avl_node_ptr_t );
 
+avl_node_ptr_t balance( avl_node_ptr_t );
+
 #endif
