+++ src/kernel/panic.c
@@ -38,7 +38,11 @@ void panic_init( void ) {
  * @param file File that invoked the panic
  * @param line Line where panic was called
  */
-void panic( const char* restrict message, const char* restrict file, uint32_t line ) {
+void __no_return panic(
+  const char* restrict message,
+  const char* restrict file,
+  uint32_t line
+) {
   // panic init
   panic_init();
