+++ src/lib/avl/Makefile.am
@@ -5,5 +5,12 @@ AM_CFLAGS = -imacros ${top_builddir}/include/config.h -I${top_srcdir}/include -I
 AM_CCASFLAGS = -imacros ${top_builddir}/include/config.h -I${top_srcdir}/include -I${top_srcdir}/include/lib
 
 noinst_LIBRARIES = libavl.a
-libavl_a_SOURCES = \
-  avl.c
+libavl_a_SOURCES =\
+  balance.c \
+  find.c \
+  insert.c \
+  max.c \
+  min.c \
+  prepare.c \
+  print.c \
+  remove.c
