+++ src/platform/rpi/serial.c
@@ -22,12 +22,12 @@
 #include <stdbool.h>
 
 #include <arch/arm/delay.h>
-#include <arch/arm/mmio.h>
 
 #include <platform/rpi/peripheral.h>
 #include <platform/rpi/gpio.h>
 #include <platform/rpi/mailbox/property.h>
 
+#include <kernel/io.h>
 #include <kernel/serial.h>
 
 /**
@@ -48,29 +48,29 @@ void serial_init( void ) {
   uint32_t base = ( uint32_t )peripheral_base_get( PERIPHERAL_GPIO );
 
   // Disable UART0.
-  mmio_write( base + UARTCR, 0 );
+  io_out32( base + UARTCR, 0 );
 
   // Setup the GPIO pin 14 && 15
-  uint32_t ra = mmio_read( base + GPFSEL1 );
+  uint32_t ra = io_in32( base + GPFSEL1 );
   ra = ( uint32_t )( ( int32_t ) ra & ~( 7 << 12 ) );
   ra |= 4 << 12;
   ra = ( uint32_t )( ( int32_t ) ra & ~( 7 << 15 ) );
   ra |= 4 << 15;
-  mmio_write( base + GPFSEL1, ra );
+  io_out32( base + GPFSEL1, ra );
 
   // Disable pull up/down for all GPIO pins & delay for 150 cycles.
-  mmio_write( base + GPPUD, 0 );
+  io_out32( base + GPPUD, 0 );
   delay( 150 );
 
   // Disable pull up/down for pin 14,15 & delay for 150 cycles.
-  mmio_write( base + GPPUDCLK0, ( 1 << 14 ) | ( 1 << 15 ) );
+  io_out32( base + GPPUDCLK0, ( 1 << 14 ) | ( 1 << 15 ) );
   delay( 150 );
 
   // Write 0 to GPPUDCLK0 to make it take effect.
-  mmio_write( base + GPPUDCLK0, 0 );
+  io_out32( base + GPPUDCLK0, 0 );
 
   // Clear pending interrupts.
-  mmio_write( base + UARTICR, 0x7FF );
+  io_out32( base + UARTICR, 0x7FF );
 
   // query mailbox to get uart clock rate
   mailbox_property_init();
@@ -92,21 +92,21 @@ void serial_init( void ) {
   uint32_t frd = ( uint32_t )( ( fractional * 64 ) + 0.5 );
 
   // write baud rate
-  mmio_write( base + UARTIBRD, brd );
-  mmio_write( base + UARTFBRD, frd );
+  io_out32( base + UARTIBRD, brd );
+  io_out32( base + UARTFBRD, frd );
 
   // Enable FIFO & 8 bit data transmissio (1 stop bit, no parity).
-  mmio_write( base + UARTLCRH, ( 1 << 4 ) | ( 1 << 5 ) | ( 1 << 6 ) );
+  io_out32( base + UARTLCRH, ( 1 << 4 ) | ( 1 << 5 ) | ( 1 << 6 ) );
 
   // Mask all interrupts.
-  mmio_write(
+  io_out32(
     base + UARTIMSC,
     ( 1 << 1 ) | ( 1 << 4 ) | ( 1 << 5 ) | ( 1 << 6 )
     | ( 1 << 7 ) | ( 1 << 8 ) | ( 1 << 9 ) | ( 1 << 10 )
   );
 
   // Enable UART0, receive & transfer part of UART.
-  mmio_write( base + UARTCR, ( 1 << 0 ) | ( 1 << 8 ) | ( 1 << 9 ) );
+  io_out32( base + UARTCR, ( 1 << 0 ) | ( 1 << 8 ) | ( 1 << 9 ) );
 
   // set flag
   serial_initialized = true;
@@ -127,8 +127,8 @@ void serial_putc( uint8_t c ) {
   uint32_t base = ( uint32_t )peripheral_base_get( PERIPHERAL_GPIO );
 
   // Wait for UART to become ready to transmit.
-  while ( 0 != ( mmio_read( base + UARTFR ) & ( 1 << 5 ) ) ) { }
-  mmio_write( base + UARTDR, ( uint32_t ) c );
+  while ( 0 != ( io_in32( base + UARTFR ) & ( 1 << 5 ) ) ) { }
+  io_out8( base + UARTDR, c );
 }
 
 /**
@@ -146,10 +146,10 @@ uint8_t serial_getc( void ) {
   uint32_t base = ( uint32_t )peripheral_base_get( PERIPHERAL_GPIO );
 
   // Wait for UART to become ready for read
-  while ( mmio_read( base + UARTFR ) & ( 1 << 4 ) ) { }
+  while ( io_in32( base + UARTFR ) & ( 1 << 4 ) ) { }
 
   // return data
-  return ( uint8_t )mmio_read( base + UARTDR );
+  return io_in8( base + UARTDR );
 }
 
 /**
@@ -165,7 +165,7 @@ void serial_flush( void ) {
   uint32_t base = ( uint32_t )peripheral_base_get( PERIPHERAL_GPIO );
 
   // read from uart until as long as something is existing to flush
-  while ( ! ( mmio_read( base + UARTFR ) & ( 1 << 4 ) ) ) {
+  while ( ! ( io_in32( base + UARTFR ) & ( 1 << 4 ) ) ) {
     serial_getc();
   }
 }
